+++ pkg/R/data.R
@@ -490,6 +490,7 @@
 #'   of population and year in time \emph{t}.}
 #'   \item{matrixqc}{A vector of integers used in the quality control section of
 #'   \code{lefkoMat} summary statements.}
+#'   \item{dataqc}{Currently a vector with two \code{NA} values.}
 #' }
 #' 
 #' @source Davison, R. et al. 2010. Demographic effects of extreme weather